+++ hmri_code_tests/hmri_test_utils.m
@@ -26,6 +26,21 @@ classdef hmri_test_utils
             % Steady state solution for dual-TR spoiled acquisition with angle in degrees
             S=hmri_test_utils.dualTRernst(deg2rad(alpha),TR1,TR2,R1);
         end
+
+        function S=HelmsMTwApprox(alpha,TR,R1,delta)
+            % Approximate steady state solution for dual flip angle GRE
+            % experiment where: 
+            %   1. the two pulses are approximately coincident
+            %   2. both flip angles alpha and beta are small, and 
+            %   3. beta is related to delta by delta = beta^2/2 ~ 1 - cos(beta).
+            % See Helms, et al. (MRM 2008 https://doi.org/10.1002/mrm.21732)
+            % for more details. Compared to Helms, et al., the amplitude of the 
+            % spoiled gradient echo under fully relaxed conditions, A, is here 
+            % implicitly 1.
+            
+            S=(alpha.*TR.*R1)./(alpha.^2/2 + delta + TR.*R1);
+            
+        end
         
         % TODO: warn about activating legacy RNG
         function seedRandomNumberGenerator